update for recording, sensor buffer added, various other updates

+++ Timeline.h
@@ -44,7 +44,7 @@ class Timeline
     Timeline();
     virtual ~Timeline();
 
-    void setSkeletonToTime(unsigned int time);
+    void setSkeletonToTime(uint64_t time);
 
     void setSkeleton(Skeleton* skeleton);
     Skeleton* getSkeleton();
@@ -54,12 +54,12 @@ class Timeline
     std::string getChannelAffiliationName(int channelId) const;
 
     // TODO(JK#1#): maybe make toChannel unsigned
-    void insert(MotionSequence* sequence, int toChannel, unsigned int time);
-    void insert(MotionSequenceChannel* channel, int toChannel, unsigned int time, std::string name = "");
-    void insert(const TimelineTrack &track, int toChannel, unsigned int time);
+    void insert(MotionSequence* sequence, int toChannel, uint64_t time);
+    void insert(MotionSequenceChannel* channel, int toChannel, uint64_t time, std::string name = "");
+    void insert(const TimelineTrack &track, int toChannel, uint64_t time);
 
     void moveChannelsDown(int startChannel, unsigned int numChannels);
-    void moveTrack(int trackId, int toChannel, unsigned int toTime);
+    void moveTrack(int trackId, int toChannel, uint64_t toTime);
 
     bool erase(int trackId);
 
@@ -70,19 +70,19 @@ class Timeline
     void sortChannels();
 
     // cut the track at given absolute time point (if it covers the time point)
-    void cut(int trackId, unsigned int time);
+    void cut(int trackId, uint64_t time);
     // cut all tracks covering the given time point at this time point
-    void cut(unsigned int time);
+    void cut(uint64_t time);
 
     TimelineTrack* getTrack(int id);
-    TimelineTrack* getTrack(int channel, unsigned int time);
+    TimelineTrack* getTrack(int channel, uint64_t time);
 
-    TimelineTrack* getTrackBefore(int channel, unsigned int time);
-    TimelineTrack* getTrackAfter(int channel, unsigned int time);
+    TimelineTrack* getTrackBefore(int channel, uint64_t time);
+    TimelineTrack* getTrackAfter(int channel, uint64_t time);
 
-    bool isBetweenTwoTracks(int channel, unsigned int time) const;
+    bool isBetweenTwoTracks(int channel, uint64_t time) const;
 
-    std::vector<TimelineTrack*> getInRange(int channel, unsigned int startTime, unsigned int endTime);
+    std::vector<TimelineTrack*> getInRange(int channel, uint64_t startTime, uint64_t endTime);
 
     std::map<int, TimelineChannel>::iterator beginChannels() { return _channels.begin(); }
     std::map<int, TimelineChannel>::iterator endChannels() { return _channels.end(); }
